final commit

+++ lab_2/Program.cs
@@ -18,16 +18,16 @@ namespace lab_2
             if (tmp1 == 1)
             {//действия с одним комплексным числом (умножение на вещественное число)
                 Console.WriteLine("Введите real часть числа");
-                int r = int.Parse(Console.ReadLine());
+                double r = double.Parse(Console.ReadLine());
                 Console.WriteLine("Введите imaginary часть числа");
-                int i = int.Parse(Console.ReadLine());
+                double i = double.Parse(Console.ReadLine());
                 ComplexNumber complex = new ComplexNumber 
                 {
                 real = r,
                 imaginary = i
                 };
                 Console.WriteLine("Введите множитель");
-                int factor = int.Parse(Console.ReadLine());
+                double factor = double.Parse(Console.ReadLine());
                 complex *= factor;
                 Console.WriteLine("z = "+complex.real+" + i*("+complex.imaginary+")");
             }
@@ -37,18 +37,18 @@ namespace lab_2
                 Console.WriteLine("2. Вычитание");
                 int tmp2 = int.Parse(Console.ReadLine());
                 Console.WriteLine("Введите real часть 1 числа");
-                int r1 = int.Parse(Console.ReadLine());
+                double r1 = double.Parse(Console.ReadLine());
                 Console.WriteLine("Введите imaginary часть 1 числа");
-                int i1 = int.Parse(Console.ReadLine());
+                double i1 = double.Parse(Console.ReadLine());
                 ComplexNumber complex1 = new ComplexNumber 
                 {
                 real = r1,
                 imaginary = i1
                 };
                 Console.WriteLine("Введите real часть 2 числа");
-                int r2 = int.Parse(Console.ReadLine());
+                double r2 = double.Parse(Console.ReadLine());
                 Console.WriteLine("Введите imaginary часть 2 числа");
-                int i2 = int.Parse(Console.ReadLine());
+                double i2 = double.Parse(Console.ReadLine());
                 ComplexNumber complex2 = new ComplexNumber 
                 {
                 real = r2,
@@ -69,9 +69,8 @@ namespace lab_2
     }
     class ComplexNumber
     {
-        public int real;
-        public int imaginary;
-        //public ComplexNumber(int r, int i){real = r; imaginary = i;}
+        public double real;
+        public double imaginary;
         public static ComplexNumber operator + (ComplexNumber comNum1, ComplexNumber comNum2)
         {
             ComplexNumber CN = new ComplexNumber
@@ -90,14 +89,14 @@ namespace lab_2
             };
             return CN;  
         }
-        public static ComplexNumber operator * (ComplexNumber comNum, int num)
+        public static ComplexNumber operator * (ComplexNumber comNum, double num)
         {
             ComplexNumber CN = new ComplexNumber
             {
                 real = comNum.real*num,
                 imaginary = comNum.imaginary*num
             };
-            return CN;  
+            return CN;
         }
     } 
 }
